dasharo-security: tpm-support.robot: add rootfs enc-decryption test
Test whether it is possible to encrypt/decrypt rootfs using TPM device.

The reason why clevis is being used here:

Ubuntu 22.04 LTS has problems regarding TPM and its collaboration with
cryptsetup as described here:
https://bugs.launchpad.net/ubuntu/+source/cryptsetup/+bug/1980018 .
The problem seems to be solved in the release of Ubuntu 23.10 as described
here:
https://ubuntu.com/blog/tpm-backed-full-disk-encryption-is-coming-to-ubuntu?

So, clevis has been chosen as an only one working tool in these conditions.

Signed-off-by: Daniil Klimuk <daniil.klimuk@3mdeb.com>