keywords.robot: list tests revision in log and report
This is done to keep track of which version of test was actually run,
as well as whether local changes were made. It should make it possible
to repeat the same tests in the future and hopefully get the same
results.

Link to source code of test is also added to suite metadata, but this
may be just an approximation in case of local modifications. In such
case, reported revision will have '-dirty' appended to the digest.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>