dasharo-compatibility/network-boot.robot: removed check for install/update
Copy of https://github.com/Dasharo/open-source-firmware-validation/issues/321:
Even though the name suggests that booting to DTS should be enough, the test
actually checked if DTS knows about Dasharo firmware for given platform. This
isn't true when doing tests before initial releases. There are dedicated tests
for DTS that check for availability of updates, and lack of them doesn't mean
that there is a problem with network boot.