lib/robot.sh: Fix faulty condition
In theory the previous conditional should work as expected,
that is clear the _robot_args variable if `--` substring
was not found. When there is no `--` seperating test suite
names from additonal args fails.
In practice the condition was always true for some reason.
Using grep is using a cannon to kill a fly, but it works.

Signed-off-by: Filip Gołaś <filip.golas@3mdeb.com>