lib/flash.robot: Fix FW_FILE not being sent to DUT via SSH
flashrom is executed on the DUT. The test gives the flash command the
name of the file on the Host machine. It does not exist on the DUT
unless we sent it there. Maybe this was working before because everytime
someone was testing a laptop they first sent the file to the DUT to
flash the new FW manually and the file stayed there?

Signed-off-by: Filip Gołaś <filip.golas@3mdeb.com>