usb: integrate and improve fixes for detect/boot suites
Tested `usb-detect.robot` and `usb-boot.robot` in two configurations:
- PiKVM (MSI Z690)
- Telnet (Odroid H4)

These tests now assume that `Dasharo Tools Suite` USB stick is inserted
to the DUT.

This is required for the `usb-boot` especially, as we assume to test
booting DTS from USB as a reference OS.

For the `usb-detect`, we may define other types of devices as well. At
this moment, `Dasharo Tools Suite` entry is the one we are looking for
in the boot menu.

Signed-off-by: Maciej Pijanowski <maciej.pijanowski@3mdeb.com>