Add functionality for running tests on a remote machine
This commit introduces a new feature that allows users to run tests on a
remote machine. This is particularly useful when running QEMU on the
local machine is not desirable due to security concerns or memory
limitations.

The changes include updates to the documentation to explain the new
feature and a new script, `remote-runner.sh`, which handles the
synchronization of a local directory to a remote machine and the
execution of QEMU based on parameters on the remote machine.

Signed-off-by: Piotr Król <piotr.krol@3mdeb.com>