scripts: refactor qemu-run.sh and remote-runner.sh scripts
This commit refactors the qemu-run.sh and remote-runner.sh scripts to
improve their flexibility and usability.

In qemu-run.sh, the ACTION parameter has been split into ACTION and
FIRMWARE parameters. This allows the user to specify the action (os or
os_install) and the firmware (uefi or seabios) separately, providing
more flexibility in the configuration of the QEMU instance. The usage
instructions have been updated accordingly. Now we are not limited to
run tests or install only on uefi.

In remote-runner.sh, the number of required parameters has been
increased from 4 to 5 to accommodate the new FIRMWARE parameter in
qemu-run.sh. The usage instructions and error messages have been updated
to reflect this change.

Additionally, the shebang in remote-runner.sh has been changed from
different systems.

Signed-off-by: Piotr Król <piotr.krol@3mdeb.com>