dasharo-compatibility/network-boot.robot: correct PXE003.01 goals
The current PXE003.01 test, which is supposed to check the functionality
of autoboot, does not perform as expected. The test passes based on the
appearance of a follow-up menu "Please select boot device", which varies
under SeaBIOS.

The issue arises from the lack of DHCP returning the necessary
information, leading to the "Nothing to boot iPXE error:
https://ipxe.org/err/2d03e1".

To correct this, DHCP should return a root-path and possibly combine
it with TFTP to retrieve the file. This approach is feasible in a QEMU
environment with customizable DHCP and TFTP. However, it might be more
challenging in a lab setting.

To fix above issue this commit change description of the test and make
it work for both uefi and seabios firmware.

Signed-off-by: Piotr Król <piotr.krol@3mdeb.com>