lib/network.robot: Change connection detection mechanism
Swap keyword cross-dependency. Previously, connection detection
would be attempted, even if our IP was not set. This makes certain
tests fail (especially when a package is installed). Furthermore,
we technically don't need Internet connection to have IP address
assigned, so we also bring this behavior to what might be expected.

Signed-off-by: Sebastian Czapla <sebastian.czapla@3mdeb.com>