lib/utc.robot: Add library with UTC tests implementation
UTC test suite is generated automatically using a script.
It is already over 3k lines in size.
The generic keywords that implement the test cases, which
(currently) repeat up to 18 times during the test suite
were moved to a library, so that the test suite can remain
fully automatically generated.

The keyword implementations are based on, and mostly just
copies of the implementations from the old versions of
the UTC test case.

Signed-off-by: Filip Gołaś <filip.golas@3mdeb.com>