dasharo-performance/fast-boot.robot: Make the test work with telnet
The test assumed that the device is already booted to an OS, which probably
is true for laptops. Add necessary Power On's and initial booting to OS
on the beginning of the test to handle a case where the serial connection
can be used to navigate and select OSes to boot.

Also do not log in to OS again after "Measure FW Boot Time On Linux",
because this keyword keeps the connection to the OS.

TEST=Run the test and PASS it on ODROID H4 Ultra

Signed-off-by: Michał Żygowski <michal.zygowski@3mdeb.com>