sleep-lib: Allow `[s2idle] deep` as a mark of s0ix enabled
That's just what Fedora shows when s0ix is enabled. It either shows
`[s2idle] deep` or `s2idle [deep]` meaning that both s2idle and deep
suspends are supported, but only one of them is selected at a given moment.
Something must have changed a bit in Dasharo if before it was
either `[s2idle] shallow` or `s2idle [deep]`, but it doesn't look
like a bug.

Signed-off-by: Filip Gołaś <filip.golas@3mdeb.com>