dasharo-compatibility/power-after-fail.robot: Fix powering on based on var state
When power after fail test are executed, the state of power after fail variable
is changed. The Power On keyword does not take it int oaccount and uses only
the default state of the variable. This results in obsolete power button presses
which often lead to powering off the platform right after PSU is powered on.
Because of that the tests fail due to power LED not being lit up.

Define another variable that is updated runtime and represents current state of
the power after faile variable.

Signed-off-by: Michał Żygowski <michal.zygowski@3mdeb.com>