Login To Linux Via SSH: Open new connection on every try
It is especially useful if switching between OSes goes wrong.
If a tester intervenes manually and switches the OS to the target
one while OSFV tries to log in, further tries will fail, because
the server's identity will change. Running Open Connection again
will allow to login to a server that has a different identity
from the one that we tried to log into at the beginning.