dts: profile: msi-pro-z[67]90*ddr5 UEFI Update - DPP: get FUM check back
The FUM check is not present in DTS v2.7.1 on wich the profiles were
originally collected. The check was added in commit
ba25f2944992f83750dd39a9240dae4911e09c0a in Dasharo/dts-scripts repo.
and will be present in the next stable DTS version.
This commit adds the FUM check back into profile to be compatible with
the DTS on develop branch in Dasharo/meta-dts.

Signed-off-by: Daniil Klimuk <daniil.klimuk@3mdeb.com>